app/requirements.txt
streamlit
pillow
requests
openai
matplotlib
azure-ai-vision
